requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
Django==4.2.6
django-paypal==2.0
idna==3.4
Pillow==10.1.0
pypng==0.20220715.0
python-decouple==3.8
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
psycopg2-binary==2.9.9
